Update patch set 3
Patch Set 3:

(2 comments)

Great Libo, I'll have a look there.

The reason for the placement is that we could have policies being received from components other than SDC, maybe other components in ONAP or ever components outside ONAP. It's easy to design for this now rather than having to do a refactor later.

We'll define an interface that all sources of policies implement and SDC will be the first to implement that interface. I'll put in a draft of that either later today or early next week.

Patch-set: 3