Update patch set 2
Patch Set 2: Code-Review+1

We might need to have a more complex structure in the distribution model in the future if the amount of "things" we are receiving and forwarding grows. We can do that when and if it arises.

Patch-set: 2
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1