Initial moduel structure for SDC handling
This submission defines the initial module structure for
this handling and add the IConfiguration implementation and
test code, use the json format as the configuration.
rename the class name to sensible one with PSSD prefix.
remove static variable to make it possible for more than one
subscription to SDC.
improve the validate func to validate each mandatory parameters
and check the int parameter's value

Change-Id: I3106ad9d4535b1251eae064e803805b92378f6f1
Issue-ID: POLICY-956
Signed-off-by: liboNet <libo.zhu@intel.com>