Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Libo, there seem to be many fixes and new code in here. The fixes for checkstyle are great. But ONAP developer best practices are for a single JIRA. So you need to break this up and submit fixes that are specific to each JIRa. One for sonar/checkstyle, one for IConfiguration and I know there is another separate JIRA for the INotificationCallback.

Patch-set: 1
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1