Update patch set 2
Patch Set 2:

Thanks, Ram for this patch.
it is always helpful and nice to do code refactor. do you think is it possible to do it after the functionality of code works?
for the status logic flow, please bear in mind, it is an asynchronous mechanism, that's the key point and the notification callback is 2nd key part. of course, the Tosca representation  is 3rd one. 

we may need to have a chat :)

Patch-set: 2