Update patch set 1
Patch Set 1:

Hello Michael, I see policyLogger.properties .. I have a question what is it using for logging .. the logging strategy used in PE side is very poor.   For example in drools we use logback directly.   Just trying to understand what logging mechanisms are used.  If we are using PolicyLogger moving in new projects, I discourage its usage, but we can work out if that's the case at a later point.

Patch-set: 1