Update patch set 2
Patch Set 2:

In theory though, we should be moving to the next version, provided that 2.0.0 has been released.   Not sure what that will be 2.1.0-SNAPSHOT (?).   I guess we have not gone through that exercise yet for any of the repos.   

It may be OK but unorthodox to push changes to 2.0.0-SNAPSHOT and then move forward the version at some point prior to release new artifacts.

What do you guys think best way to move forward?

Patch-set: 2