Update patch set 1
Patch Set 1: Code-Review-2

Holding this for a separate review to come through that fixes sonar/checkstyle issues. Unfortunately there was an oversight on that code. We will rebase once that commit is merged.
Lianho, also we are requiring JUnit test coverage for all new code so we don't fall out of range for Release Requirements.

Patch-set: 1
Label: Code-Review=-2