Update patch set 1
Patch Set 1:

Before submitting this review, the 3 from Lianho have to be cherry picked from master. Then the review I just submitted this morning to fix all the checkstyle and 2 sonar bugs/criticals. Then we rebase and determine which of the reviews from Lianho and Libho should go next.

Patch-set: 1