Fix unchecked cast sonar issue
The variable is currently unused as flagged by sonar. But
the real issue is the @SuppressWarnings("unchecked") should
not be used unless there is a check to ensure that the
object you are casting to really is that object.

So added the check for that which should remove the
original sonar issues.

Issue-ID: POLICY-1256
Change-Id: Iad7dbbf02a4dd48648fc9d5c20595c9f0f8d7acb
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>