Update patch set 1
Patch Set 1:

Technically not needed as the root LICENSE is in place. Not Java so the ONAP License check doesn't care.

Since we merged the other branch we can leave this as is for this review. If Lianhao wants to add a license he can do so.

Patch-set: 1
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0