Update patch set 3
Patch Set 3:

seems that we would be better off with a separate model for these HPA placement policies rather than hard coding it into the software. Anytime anything changes in the model, then this artifact has to change which is not sustainable. We will need to re-visit this in Dublin.

Patch-set: 3