Update patch set 3
Patch Set 3:

> Thinking ahead about recovery. If a subset of policies succeed, and
 > a subset fails. Possibly re-distribution can be done via the SDC
 > GUI. It could be done in any circumstance. Do we think we are able
 > to handle that very likely situation? Any past failed policies may
 > *possibly* work again? As well as any policies that were previously
 > loaded as part of success scenario, would gently error out?

As per current code, when we get a list of policies we create/deploy them one by one and if one of the policy fails we go to next one in the loop. After completing the list, we check if any of the policy has failed and raise exception if needed. 
Now if distribution happens again then we will try to create/deploy the full list again from distribution side but the duplication is already handled by API, PAP & PDP's.

Patch-set: 3
Label: Code-Review=0