Update patch set 6
Patch Set 6:

After discussion with the team, we decided to modify the StandardCoder to disable html escaping.  Once that's done, this should work using "=".

In the meantime, you can wait until that's done, or you can change to use StandardCoder and change the "=" as mentioned in my previous reply.  I'll leave that up to you.

Patch-set: 6
Label: Code-Review=0