Use new RestClientParameters class in distribution
This is a re-do of a previously reverted review.  It's mostly the same
as before, but there are some minor differences.

Replaced LifecycleApiForwarderParameters with RestClientParameters.
Also modified the code to get the HttpClient objects once.
Also removed some unused json test files.

This requires changes to property files in CSITs and OOM:
- add "clientName" to the api & pap structures
- rename "hostName" to "hostname"
- rename "isHttps" to "useHttps" and move it inside the api & pap
  structures

Issue-ID: POLICY-3456
Change-Id: I32b04bc3b6880ec7798814dad268143bc3b06faf
Signed-off-by: Jim Hahn <jrh3@att.com>